requirements.txt
aiohttp==3.8.3
discord.py==2.0.1
requests==2.27.1
